export interfaces module use
