+++ src/components/hooks/useApplicationData.tsx
@@ -1,21 +1,6 @@
 import { useEffect, useState } from "react"
 import axios from "axios"
-
-export interface Group {
-  id: number,
-  name: string
-}
-export interface Post {
-  id: number,
-  group_id: number,
-  user_id: number,
-  created_at: number
-}
-export interface User {
-  group: number;
-  groups?: Group[];
-  posts?: Post[];
-}
+import {User} from "../../helpers/interfaces"
 
 export default function useApplicationData() {
